% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEV_auth.R
\name{DEV_auth}
\alias{DEV_auth}
\alias{DEV_auth_set}
\alias{DEV_auth_get}
\title{Configure Google Sheets Authentication}
\usage{
DEV_auth_set(email, file = ".secrets/gs4-token.rds")

DEV_auth_get()
}
\description{
\itemize{
\item Use \code{DEV_auth_set()} to create an encrypted gs4 token.
\item Use \code{DEV_auth_get()} to read the token.
}
}
