% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyReviewer.R
\name{ShinyReviewer}
\alias{ShinyReviewer}
\title{ShinyReviewer R6 Class.}
\description{
ShinyReviewer R6 Class.

ShinyReviewer R6 Class.
}
\details{
Use this to create ShinyReviewer application instances.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ShinyReviewer-new}{\code{ShinyReviewer$new()}}
\item \href{#method-ShinyReviewer-run_app}{\code{ShinyReviewer$run_app()}}
\item \href{#method-ShinyReviewer-clone}{\code{ShinyReviewer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyReviewer-new"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyReviewer-new}{}}}
\subsection{Method \code{new()}}{
Create a ShinyReviewer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyReviewer$new(reviews, content_meta, group_meta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reviews}}{A data frame with \code{content_id}, \code{timestamp}, \code{rating},
\code{difficulty}, \code{blacklist}, and \code{notes}. The row with the most recent
\code{timestamp} is used for each \code{content_id}.}

\item{\code{content_meta}}{A data frame with \code{group_id} (optional), \code{id}, and
\code{title}. If \code{group_id} is provided, content is grouped within cards.}

\item{\code{group_meta}}{(Optional) A df giving \code{id}, \code{title}, and \code{img_src},
used to create group cards.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyReviewer-run_app"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyReviewer-run_app}{}}}
\subsection{Method \code{run_app()}}{
Run the ShinyReviewer application.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyReviewer$run_app(title, callback = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{An application title.}

\item{\code{callback}}{(Optional) A \verb{function(df)} to do something with the
reviews submitted within a session when it ends.}

\item{\code{...}}{Options passed to \code{shiny::shinyApp()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyReviewer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyReviewer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyReviewer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
